+++ index.html
@@ -43,6 +43,6 @@
       </section>
     </main>
     <footer><p>Copyright ...</p></footer>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
